build: split the test suite into two and run browserify tests separately
This is because otherwise we load zone.js twice (once regularly and once via browserify),
which results in double patching and all kinds of bad things.

In the future we should consider having a way to singnal that Zone should not init when loaded
or make it possible to cleanly uninstall Zone.

Closes #86