fix: don't fork new zones for callbacks from the root zone
BREAKING CHANGE: New child zones are now created only from a async task that installed a custom zone.

Previously even without a custom zone installed (e.g. LongStacktracesZone), we would spawn new
child zones for all asynchronous events. This is undesirable and generally not useful.

It does not make sense for us to create new zones for callbacks from the root zone since we care
only about callbacks from installed custom zones. This reduces the overhead of zones.

This primarily means that LongStackTraces zone won't be able to trace events back to Zone.init(),
but instead the starting point will be the installation of the LongStacktracesZone. In all practical
situations this should be sufficient.

Closes #92