fix(zone): fix conflict with Polymer elements
I noticed when including Polymer 1.0 elements in an Angular 2 app, this line throws an error saying it can't find `value` of undefined. I've added a simple workaround which seems to fix the issue and (so far at least) doesn't seem to cause any additional issues with Polymer or Angular.

cc @justinfagnani @addyosmani