chore: clear stack trace on async tests
When you have in async test (test with `done` argument) jasmine will 
execute the next test synchronously in the done handle. This makes sense
for most tests, but now with zones. With zones running next test 
synchronously means that the current zone does not get cleared. This 
results in a chain of nested zones, which makes it hard to reason about 
it. We override the `clearStack` method which forces jasmine to always 
drain the stack before next test gets executed.