zone.js does not support timeouts that are passed as strings for eval.
zone.js patches to window.setTimeout and window.setInterval do not allow strings containing javascript to be eval'd.

Using zone.js in conjunction with a large legacy codebase requires us to allow eval strings be passed to timeout calls.  Currently, as zone.js expects the value to be a function, strings meant for eval fail.  This edit resolves that issue.