fix: stack rewriting now works with source maps
In Chrome for the source maps to work the object must be an instance of
native `Error`. This means that we can’t return a subclass of `Error`
or error handling in the dev console will not work properly.

In addition the stack frames must have a certain format or the source
mapping is disabled. For this reason we have changed the long stack
format to conform to that shape.