fix: add support for subclassing of Errors
Calling `new Error` returns a new instance of `Error`. This means
that subclassing of `Error` needs to take that into account as calling
`super` will not expect return value. 

To fix this `ZoneAwareError` detects to see if it was called with `new`
and if so will copy the fields from the `Error` to the current instance.