fix(REVERT): remove zone internal stack frames in error.stack (#632)
This feature was reverted because it was causing a performance regression.
The setup of all of the stack traces added about 20-25ms to the total
time of the Zone initialization. This was a bit more than 50% of overall
time zone needed to initialize. While this is a useful feature, the
performance cost does not justify its cost.