fix(spec): do not count requestAnimationFrame as a pending timer
We added a new fature to FakeAsyncTestSpec to treat a
requestAnimationFrame as a timer with 16ms. However this breaks existing
Angular project tests because the RAF is reported as a pending timer at
the end of the test.

This change makes sure that existing Angular tests are not broken by not
counting RAF-s as pending timers at the end of a test.