Add `unpkg` field to package.json
This change adds an `unpkg` field to package.json that will allow users
to load zone.js via unpkg w/out using the ?main=browser query parameter.

Also, we're removing support for using the `browser` fallback on
unpkg.com due to the fact that it was originally intended to be used by
bundlers, which unpkg is not.

For more info, please see https://github.com/unpkg/unpkg/issues/63