fix: ensure that EventTarget is patched prior to legacy property descriptor patch (#1214)
If the event that EventTarget is defined on a device that also does not support patch via property descriptor, the result is that the XMLHttpRequest class would be patched prior to the event target being patched (due to returning in event-target-legacy if EventTarget is defined). This results in the wrapper that replaces XMLHttpRequest missing the patched listener fields. E.g. __zone_symbol__addEventListener and __zone_symbol__removeEventListener
event-target-legacy.

The result is that for any attempted http requests, a type error occurs in scheduleTask in browse.ts -> patchXhr, due to attempting call to non-existent __zone_symbol__addEventListener and __zone_symbol__removeEventListener, causing the requests to fail.

Ensuring that patching of event target for legacy path takes place before legacy property descriptor patch eliminates this problem.